+++ Engine/Source/Runtime/Core/Private/HAL/UnrealMemory.cpp
@@ -374,7 +374,7 @@ static int FMemory_GCreateMalloc_ThreadUnsafe()
 // On Mac it's too early to log here in some cases. For example GMalloc may be created during initialization of a third party dylib on load, before CoreFoundation is initialized
 #if !PLATFORM_DESKTOP
 	// by this point, we assume we can log
-	double SizeInMb = ProgramSize / (1024.0 * 1024.0);
+	double SizeInMb = (double)ProgramSize / (1024.0 * 1024.0);
 	FPlatformMisc::LowLevelOutputDebugStringf(TEXT("Used memory before allocating anything was %.2fMB\n"), SizeInMb);
 	UE_LOG(LogMemory, Display, TEXT("Used memory before allocating anything was %.2fMB"), SizeInMb);
 #endif
