+++ Engine/Source/Runtime/Core/Private/Stats/StatsData.cpp
@@ -1101,9 +1101,9 @@ void FStatsThreadState::UpdateStatMessagesMemoryUsage()
 		const double InvMB = 1.0f / 1024.0f / 1024.0f;
 
 		// Format lines to be displayed on the hud.
-		const FString Current = FString::Printf( TEXT("Current: %.1f"), InvMB*CurrentNumStatMessages*sizeof(FStatMessage) );
-		const FString Max = FString::Printf( TEXT("Max: %.1f"), InvMB*(int64)MaxNumStatMessages*sizeof(FStatMessage) );
-		const FString Total = FString::Printf( TEXT("Total: %.1f") , InvMB*TotalNumStatMessages*sizeof(FStatMessage));
+		const FString Current = FString::Printf( TEXT("Current: %.1f"), float(InvMB*double(CurrentNumStatMessages*sizeof(FStatMessage))) );
+		const FString Max = FString::Printf( TEXT("Max: %.1f"), float(InvMB*double((int64)MaxNumStatMessages*sizeof(FStatMessage))) );
+		const FString Total = FString::Printf( TEXT("Total: %.1f") , float(InvMB*double(TotalNumStatMessages*sizeof(FStatMessage))) );
 
 		UE_LOG(LogStats, Verbose, TEXT("%s, %s, %s"), *Current, *Max, *Total );
